Initial basic commit
This first commit adds basic components.

It adds an index.html file which contains a canvas to hold the world.
It also includes the scripts, and bootstraps the world and engine.

This commit also adds a few js modules which link together.

The world encapsulates the entities and activity of the project.
The engine takes in the world, creates a viewport to let us see into it,
and then ticks along to update() and render() the world in each moment
in time.

The engine also routes user input from the mouse and keyboard on
the canvas to the world.

The engine also creates an FPS meter for debugging purposes.