Enable composition of entities
Instead of creating entities as specific versions of other entities
(i.e., DriveableEntity is a MoveableEntity which in turn is an Entity)
we make it so that entties can have a default extention as before,
but they can now also have the flexibility to be an extention of
any other object.

Prior to this:
They were created with .create() which built upon a hard-coded
parent entity.

Now:
They can still be creatd with a .create() to build upon a hard-coded
default, but now under the hood, .create() calls the new .extend().
The .extend() can take in the default made by .create(), or another
entity supplied by the user.