Add entities to the world via a 'pending' variable
This is because we garbage collect dead entities and remove them from the 'entities'
variable of the World on each update. That garbage collection occurs by mapping and
filtering over the existing data. This made it impossible to spawn new entities
within an update loop.

This fixes that by having the "add_" methods of the world add to a pending queue,
which is concatinated into the entities variable during mapping.

Also some updates to shooting so you can hold down the trigger, boom boom