Major update: Add layers to the world
This is for parallax layering of views.

This change mostly takes the behavior from the World object, and copies
it into Layer objects, which the world can have one-or-many of.

Rendering happens from low depth to high depth in the layers.

Changes were made to rendering and viewports to take advantage of
the depth information to adjust the location of objects.

Test index file was updated to add some background tests.