Create phases for the game
This removes a lot of the per-level logic out of just being programatic
in the index file, and pushes them into "phase" objects that can be
loaded into the "game" engine.

These phases define a function to load assets into the engine as needed,
a second function to setup the entities in the world, and provides the
ability to define user-input handling functions.

This makes it possible to switch between "phases" of the game. For example,
we can load a firs phase that acts as a title screen, and then set the
handle_mouse_button method to switch us into the level-1 phase.