Modify the position of objects on different layers
This modifies the positions to be more correct such that if the
player is over a set of coordinates, all objects at that same
coordinate value on different layers are lined up.

Prior to this commit, objects at the same position would only be
lined up if they were at 0,0. This made it difficult to build levels
as coords were relative to the origin and all maps had their focus
center around that, even when the player moved away from the origin.

With this change, the world appears to the user as if the origin is
wherever the center of the viewport is.