Decouple specialized entity updating
Instead of having the tracking of the mouse be hardcoded, it's moved
into a custom function of the scene. This is because we might not
always _want_ player entities to track the mouse.

By adding a new update function to the scene, we can do special and
custom things in different scenes and different games.

Also some code cleanup and function renames tacked on.